+++ integrations/opensearch/src/haystack_integrations/components/retrievers/opensearch/embedding_retriever.py
@@ -8,6 +8,7 @@ from haystack import component, default_from_dict, default_to_dict
 from haystack.dataclasses import Document
 from haystack.document_stores.types import FilterPolicy
 from haystack.document_stores.types.filter_policy import apply_filter_policy
+
 from haystack_integrations.document_stores.opensearch import OpenSearchDocumentStore
 
 logger = logging.getLogger(__name__)
@@ -70,8 +71,16 @@ class OpenSearchEmbeddingRetriever:
         For this `custom_query`, an example `run()` could be:
 
         ```python
-        retriever.run(query_embedding=embedding,
-                        filters={"years": ["2019"], "quarters": ["Q1", "Q2"]})
+        retriever.run(
+            query_embedding=embedding,
+            filters={
+                "operator": "AND",
+                "conditions": [
+                    {"field": "meta.years", "operator": "==", "value": "2019"},
+                    {"field": "meta.quarters", "operator": "in", "value": ["Q1", "Q2"]},
+                ],
+            },
+        )
         ```
         :param raise_on_failure:
             If `True`, raises an exception if the API call fails.
@@ -175,8 +184,16 @@ class OpenSearchEmbeddingRetriever:
         For this `custom_query`, an example `run()` could be:
 
         ```python
-        retriever.run(query_embedding=embedding,
-                        filters={"years": ["2019"], "quarters": ["Q1", "Q2"]})
+        retriever.run(
+            query_embedding=embedding,
+            filters={
+                "operator": "AND",
+                "conditions": [
+                    {"field": "meta.years", "operator": "==", "value": "2019"},
+                    {"field": "meta.quarters", "operator": "in", "value": ["Q1", "Q2"]},
+                ],
+            },
+        )
         ```
 
         :returns: