+++ integrations/amazon_bedrock/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [unreleased]
+
+### 🐛 Bug Fixes
+
+- *(Bedrock)* Allow tools kwargs for AWS Bedrock Claude model (#976)
+
+### 🚜 Refactor
+
+- Remove usage of deprecated `ChatMessage.to_openai_format` (#1007)
+
+## [integrations/amazon_bedrock-v1.0.1] - 2024-08-19
+
+### 🚀 Features
+
+- Make truncation optional for bedrock chat generator (#967)
+
+### 🐛 Bug Fixes
+
+- Normalising ChatGenerators output (#973)
+
 ## [integrations/amazon_bedrock-v1.0.0] - 2024-08-12
 
 ### 🚜 Refactor